requirements.txt
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
browsermob-proxy==0.7.1
certifi==2016.2.28
cycler==0.10.0
dask==0.8.0
decorator==4.0.9
functools32==3.2.3.post2
gevent==1.1.0
greenlet==0.4.9
hashlib==20081119
IPy==0.83
ipykernel==4.3.1
ipython==4.1.2
ipython-genutils==0.1.0
Jinja2==2.8
jsbeautifier==1.6.2
jsonschema==2.5.1
jupyter-client==4.2.1
jupyter-core==4.1.0
lxml==3.5.0
MarkupSafe==0.23
matplotlib==1.5.1
mistune==0.7.2
nbconvert==4.1.0
nbformat==4.0.1
networkx==1.11
notebook==4.1.0
numpy==1.10.4
pandas==0.17.1
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
Pillow==3.1.1
ptyprocess==0.5.1
Pygments==2.1.3
pylab==0.1.3
pyparsing==2.1.0
python-dateutil==2.5.0
python-geoip==1.2
python-geoip-geolite2==2015.303
pytz==2015.7
pyzmq==15.2.0
requests==2.9.1
scikit-image==0.12.3
scikit-learn==0.17.1
scipy==0.17.0
seaborn==0.7.0
selenium==2.52.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
sympy==0.7.6.1
terminado==0.6
tld==0.7.6
toolz==0.7.4
tornado==4.3
traitlets==4.1.0
